Create Link to Internal Sections of a Page with Anchor Elements
Link to Internal Sections of a Page with Anchor ElementsPassed
a (anchor) elements can also be used to create internal links to jump to different sections within a webpage.

To create an internal link, you assign a link's href attribute to a hash symbol # plus the value of the id attribute for the element that you want to internally link to, usually further down the page. You then need to add the same id attribute to the element you are linking to. An id is an attribute that uniquely describes an element.

Below is an example of an internal anchor link and its target element:

<a href="#contacts-header">Contacts</a>
...
<h2 id="contacts-header">Contacts</h2>
When users click the Contacts link, they'll be taken to the section of the webpage with the Contacts heading element.

Change your external link to an internal link by changing the href attribute to "#footer" and the text from cat photos to Jump to Bottom.

Remove the target="_blank" attribute from the anchor tag since this causes the linked document to open in a new window tab.

Then add an id attribute with a value of footer to the <footer> element at the bottom of the page.